+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## Change Log (we do our best to adhere to [semantic versioning](http://semver.org/))
 
-#### Version: 2.0.0-RC3-SNAPSHOT
+
+#### Version: 2.0.0-SNAPSHOT
+
+
+#### Version: 2.0.0-RC4 - 2016-06-21
 - **BREAKING CHANGES**
   - World's systems array is now typed with `BaseSystem`; this should only have
     implication for custom implementations of `SystemInvocationStrategy`.